+++ src/utilities/dom/range-rect-calculator.ts
